# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LessonSourceType = typeof LessonSourceType[keyof typeof LessonSourceType];


export const LessonSourceType = {
  text: 'text',
  youtube: 'youtube',
  audio: 'audio',
  epub: 'epub',
  url: 'url',
} as const;

export type LessonStatus = typeof LessonStatus[keyof typeof LessonStatus];


export const LessonStatus = {
  saved: 'saved',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Lesson {
  id: string;
  title: string;
  /** @nullable */
  sourceUrl?: string | null;
  sourceType: LessonSourceType;
  rawText: string;
  /**
     * JSON-encoded token array
     * @nullable
     */
  tokens?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  collectionId?: string | null;
  status: LessonStatus;
  wordCount: number;
  createdAt: string;
}

export type LessonInputSourceType = typeof LessonInputSourceType[keyof typeof LessonInputSourceType];


export const LessonInputSourceType = {
  text: 'text',
  youtube: 'youtube',
  audio: 'audio',
  epub: 'epub',
  url: 'url',
} as const;

export interface LessonInput {
  title: string;
  /** @nullable */
  sourceUrl?: string | null;
  sourceType: LessonInputSourceType;
  rawText: string;
  /** @nullable */
  tokens?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  collectionId?: string | null;
  wordCount?: number;
}

/**
 * @nullable
 */
export type LessonUpdateStatus = typeof LessonUpdateStatus[keyof typeof LessonUpdateStatus] | null;


export const LessonUpdateStatus = {
  saved: 'saved',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface LessonUpdate {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: LessonUpdateStatus;
  /** @nullable */
  collectionId?: string | null;
}

export interface LessonStats {
  lessonId: string;
  totalTokens: number;
  knownCount: number;
  learningCount: number;
  unknownCount: number;
  percentKnown: number;
}

export interface VocabEntry {
  id: string;
  surface: string;
  dictionaryForm: string;
  reading: string;
  /** 0=unknown, 1-4=learning stages, 5=known, 99=ignored */
  status: number;
  timesSeen: number;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  sourceLessonId?: string | null;
  /** @nullable */
  nextReview?: string | null;
  /** @nullable */
  easeFactor?: number | null;
  /** @nullable */
  intervalDays?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface VocabUpdate {
  /** @nullable */
  status?: number | null;
  /** @nullable */
  note?: string | null;
}

export interface VocabUpsertInput {
  surface: string;
  dictionaryForm: string;
  reading: string;
  /** @nullable */
  status?: number | null;
  /** @nullable */
  sourceLessonId?: string | null;
}

export interface VocabLookupInput {
  dictionaryForms: string[];
}

export interface VocabSummary {
  total: number;
  known: number;
  learning: number;
  unknown: number;
  ignored: number;
}

export interface DictionarySense {
  glosses: string[];
  partOfSpeech?: string[];
  examples?: string[];
}

export interface DictionaryEntry {
  headword: string;
  readings: string[];
  senses: DictionarySense[];
  /** @nullable */
  jlptLevel?: string | null;
  partOfSpeech?: string[];
}

export interface Collection {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  lessonCount?: number;
  createdAt: string;
}

export interface CollectionInput {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface CollectionUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface SrsCard {
  vocabEntry: VocabEntry;
  contextSentence: string;
  /** @nullable */
  lessonTitle?: string | null;
}

export interface SrsReviewInput {
  vocabId: string;
  /** SM-2 quality rating 0-5 */
  quality: number;
}

export interface DashboardStats {
  totalKnownWords: number;
  totalWordsRead: number;
  totalListeningMinutes: number;
  totalStudyMinutes: number;
  totalLessons: number;
  lessonsInProgress: number;
  lessonsCompleted: number;
  dueForReview: number;
}

export interface DailyStats {
  date: string;
  wordsRead: number;
  listeningMinutes: number;
  newLingqs: number;
  knownWordsTotal: number;
  studyMinutes: number;
}

export interface StatsSessionInput {
  /** @nullable */
  wordsRead?: number | null;
  /** @nullable */
  listeningMinutes?: number | null;
  /** @nullable */
  newLingqs?: number | null;
  /** @nullable */
  studyMinutes?: number | null;
}

export interface ImportTextInput {
  title: string;
  text: string;
  /** @nullable */
  collectionId?: string | null;
}

export interface ImportUrlInput {
  url: string;
  /** @nullable */
  collectionId?: string | null;
}

export interface ImportYoutubeInput {
  url: string;
  /** @nullable */
  collectionId?: string | null;
}

export interface AiTranslateInput {
  sentence: string;
  /** @nullable */
  targetWord?: string | null;
}

export interface AiTranslateResult {
  translation: string;
  /** @nullable */
  wordExplanation?: string | null;
}

export type GetLessonsParams = {
/**
 * @nullable
 */
collectionId?: string | null;
/**
 * @nullable
 */
status?: GetLessonsStatus;
/**
 * @nullable
 */
search?: string | null;
};

export type GetLessonsStatus = typeof GetLessonsStatus[keyof typeof GetLessonsStatus] | null;


export const GetLessonsStatus = {
  saved: 'saved',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type GetVocabParams = {
/**
 * @nullable
 */
status?: number | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};

export type LookupVocab200 = {[key: string]: VocabEntry};

export type DictionaryLookupParams = {
word: string;
/**
 * @nullable
 */
reading?: string | null;
};

export type GetStatsHistoryParams = {
/**
 * @nullable
 */
days?: number | null;
};