# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LinguaReader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lessons:
    get:
      operationId: getLessons
      tags: [lessons]
      summary: List all lessons
      parameters:
        - in: query
          name: collectionId
          schema:
            type: ["string", "null"]
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: ["saved", "in_progress", "completed", null]
        - in: query
          name: search
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a new lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a single lesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson metadata or status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /lessons/{id}/stats:
    get:
      operationId: getLessonStats
      tags: [lessons]
      summary: Get per-lesson vocabulary statistics
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson vocabulary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonStats"

  # ─── VOCAB ──────────────────────────────────────────────────────────────
...
  /collections:
    get:
      operationId: getCollections
      tags: [collections]
      summary: List all collections
      responses:
        "200":
          description: Collections list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── SRS ────────────────────────────────────────────────────────────────
...
  /import/text:
    post:
      operationId: importText
      tags: [import]
      summary: Import raw Japanese text as a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportTextInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /import/url:
    post:
      operationId: importUrl
      tags: [import]
      summary: Import a web article by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportUrlInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /import/youtube:
    post:
      operationId: importYoutube
      tags: [import]
      summary: Import YouTube video subtitles as a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportYoutubeInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

  # ─── AI ─────────────────────────────────────────────────────────────────
  /ai/translate:
    post:
      operationId: aiTranslate
      tags: [ai]
      summary: Get AI translation for a sentence or word in context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiTranslateInput"
      responses:
        "200":
          description: AI translation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTranslateResult"
components:
  schemas:
...
    Lesson:
      type: object
      required: [id, title, sourceType, rawText, status, wordCount, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        sourceUrl:
          type: ["string", "null"]
        sourceType:
          type: string
          enum: ["text", "youtube", "audio", "epub", "url"]
        rawText:
          type: string
        tokens:
          type: ["string", "null"]
          description: JSON-encoded token array
        audioUrl:
          type: ["string", "null"]
        collectionId:
          type: ["string", "null"]
        status:
          type: string
          enum: ["saved", "in_progress", "completed"]
        wordCount:
          type: integer
        createdAt:
          type: string
    LessonInput:
      type: object
      required: [title, sourceType, rawText]
      properties:
        title:
          type: string
        sourceUrl:
          type: ["string", "null"]
        sourceType:
          type: string
          enum: ["text", "youtube", "audio", "epub", "url"]
        rawText:
          type: string
        tokens:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        collectionId:
          type: ["string", "null"]
        wordCount:
          type: integer
    LessonUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: ["saved", "in_progress", "completed", null]
        collectionId:
          type: ["string", "null"]
    LessonStats:
      type: object
      required: [lessonId, totalTokens, knownCount, learningCount, unknownCount, percentKnown]
      properties:
        lessonId:
          type: string
        totalTokens:
          type: integer
        knownCount:
          type: integer
        learningCount:
          type: integer
        unknownCount:
          type: integer
        percentKnown:
          type: number

    # ── Vocab ────────────────────────────────────────────────────────────
...
    Collection:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        lessonCount:
          type: integer
        createdAt:
          type: string
    CollectionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    CollectionUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]

    # ── SRS ──────────────────────────────────────────────────────────────
...
    ImportTextInput:
      type: object
      required: [title, text]
      properties:
        title:
          type: string
        text:
          type: string
        collectionId:
          type: ["string", "null"]
    ImportUrlInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        collectionId:
          type: ["string", "null"]
    ImportYoutubeInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        collectionId:
          type: ["string", "null"]

    # ── AI ───────────────────────────────────────────────────────────────
    AiTranslateInput:
      type: object
      required: [sentence]
      properties:
        sentence:
          type: string
        targetWord:
          type: ["string", "null"]
    AiTranslateResult:
      type: object
      required: [translation]
      properties:
        translation:
          type: string
        wordExplanation:
          type: ["string", "null"]