# lib/api-zod/src/generated/types/aiTranslateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */

export interface AiTranslateInput {
  sentence: string;
  /** @nullable */
  targetWord?: string | null;
}