# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  lessonCount?: number;
  createdAt: string;
}