# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionInput {
  name: string;
  /** @nullable */
  description?: string | null;
}