# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LinguaReader API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonInputSourceType } from './lessonInputSourceType';

export interface LessonInput {
  title: string;
  /** @nullable */
  sourceUrl?: string | null;
  sourceType: LessonInputSourceType;
  rawText: string;
  /** @nullable */
  tokens?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  collectionId?: string | null;
  wordCount?: number;
}